Documents/Materi smt 7/Algoritma Deep Learning/tugas-deep-learning-2/requirements.txt
ultralytics==8.2.0
Pillow==9.5.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
